requirements.txt
# Core Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
numpy==1.24.3

# NLP & Embeddings
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.33.0
spacy==3.7.0

# Machine Learning
scikit-learn==1.3.0
torch==2.0.1
imbalanced-learn==0.11.0  # For SMOTE

# Graph & Network Analysis
networkx==3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Fast Similarity Search
faiss-cpu==1.7.4  # Use faiss-gpu if you have CUDA

# Optimization Algorithms
pyswarms==1.3.0  # For Firework Algorithm

# Configuration Management
pyyaml==6.0.1

# Web UI
streamlit==1.28.0
gradio==4.0.0  # Alternative UI option

# Evaluation & Metrics
scikit-metrics==0.1.0

# RAG & Explanations (Optional but Impressive)
langchain==0.0.340
openai==1.3.0  # If using GPT for explanations

# API Development (Optional)
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
wikipedia==1.4.0

# Data Augmentation
nlpaug==1.1.11
googletrans==4.0.0rc1  # For back-translation

# Additional for Phase 2
seaborn==0.12.2  # Statistical visualizations

# Experiment Tracking (Optional but Professional)
# wandb==0.16.0
# mlflow==2.8.1